+++ src/containers/ConfirmationContainer/ConfirmationContainer.vue
@@ -6,7 +6,7 @@
       :signed-tx="signedTx"
       :fee="transactionFee"
       :is-hardware-wallet="isHardwareWallet"
-      :gas-price="$store.state.gasPrice"
+      :gas-price="gasPrice"
       :from="fromAddress"
       :to="toAddress"
       :value="amount"
@@ -23,7 +23,7 @@
       :signed-tx="signedTx"
       :fee="transactionFee"
       :is-hardware-wallet="isHardwareWallet"
-      :gas-price="$store.state.gasPrice"
+      :gas-price="gasPrice"
       :from="fromAddress"
       :to="toAddress"
       :value="amount"
@@ -53,6 +53,7 @@ import ConfirmModal from './components/ConfirmModal';
 import ConfirmCollectionModal from './components/ConfirmCollectionModal';
 import SuccessModal from './components/SuccessModal';
 import ConfirmSignModal from './components/ConfirmSignModal';
+import { mapGetters } from 'vuex';
 
 export default {
   components: {
@@ -84,7 +85,7 @@ export default {
       nonce: 0,
       gasLimit: 21000,
       data: '0x',
-      gasAmount: this.$store.state.gasPrice,
+      gasAmount: this.gasPrice,
       parsedBalance: 0,
       toAddress: '',
       transactionFee: 0,
@@ -105,9 +106,14 @@ export default {
     };
   },
   computed: {
+    ...mapGetters({
+      gasPrice: 'gasPrice',
+      wallet: 'wallet',
+      web3: 'web3'
+    }),
     fromAddress() {
-      if (this.$store.state.wallet) {
-        return this.$store.state.wallet.getChecksumAddressString();
+      if (this.wallet) {
+        return this.wallet.getChecksumAddressString();
       }
     }
   },
@@ -119,7 +125,7 @@ export default {
         'Transaction Hash'
       ]);
       const pollReceipt = setInterval(() => {
-        this.$store.state.web3.eth.getTransactionReceipt(newVal).then(res => {
+        this.web3.eth.getTransactionReceipt(newVal).then(res => {
           if (res !== null) {
             this.web3WalletRes = res;
             this.showSuccessModal('Transaction sent!', 'Okay');
@@ -147,10 +153,10 @@ export default {
       if (tx.hasOwnProperty('ensObj')) {
         delete tx['ensObj'];
       }
-      this.isHardwareWallet = this.$store.state.wallet.isHardware;
+      this.isHardwareWallet = this.wallet.isHardware;
       this.responseFunction = resolve;
       this.successMessage = 'Sending Transaction';
-      this.$store.state.wallet.signTransaction(tx).then(_response => {
+      this.wallet.signTransaction(tx).then(_response => {
         this.signedTxObject = _response;
         this.signedTx = this.signedTxObject.rawTransaction;
       });
@@ -161,7 +167,7 @@ export default {
       this.parseRawTx(tx);
       this.responseFunction = resolve;
       this.successMessage = 'Sending Transaction';
-      this.$store.state.wallet.signTransaction(tx).then(_response => {
+      this.wallet.signTransaction(tx).then(_response => {
         this.web3WalletHash = _response;
       });
       this.showSuccessModal(
@@ -174,7 +180,7 @@ export default {
       const newArr = [];
       this.isHardwareWallet = isHardware;
       for (let i = 0; i < tx.length; i++) {
-        this.$store.state.wallet.signTransaction(tx[i]).then(_response => {
+        this.wallet.signTransaction(tx[i]).then(_response => {
           newArr.push(_response);
         });
       }
@@ -185,7 +191,7 @@ export default {
     this.$eventHub.$on('showMessageConfirmModal', (data, resolve) => {
       this.responseFunction = resolve;
       this.messageToSign = data;
-      this.$store.state.wallet.signMessage(data).then(_response => {
+      this.wallet.signMessage(data).then(_response => {
         this.signedMessage = '0x' + _response.toString('hex');
       });
       this.signConfirmationModalOpen();
@@ -269,7 +275,7 @@ export default {
       ]);
 
       const pollReceipt = setInterval(() => {
-        this.$store.state.web3.eth.getTransactionReceipt(response).then(res => {
+        this.web3.eth.getTransactionReceipt(response).then(res => {
           if (res !== null) {
             this.$store.dispatch('addNotification', [
               this.fromAddress,
@@ -283,7 +289,7 @@ export default {
       }, 500);
     },
     async sendBatchTransactions() {
-      const web3 = this.$store.state.web3;
+      const web3 = this.web3;
       const batch = new web3.eth.BatchRequest();
       for (let i = 0; i < this.signedArray.length; i++) {
         batch.add(
@@ -310,7 +316,7 @@ export default {
       this.nonce = 0;
       this.gasLimit = 21000;
       this.data = '0x';
-      this.gasAmount = this.$store.state.gasPrice;
+      this.gasAmount = this.gasPrice;
       this.parsedBalance = 0;
       this.toAddress = '';
       this.transactionFee = 0;